fix(supplier): give the completion-approval outcome writes real transactions
Five of the nine real java:S6809 sites from the phase 3.1 triage.

WorkorderCompletionApprover's @Scheduled approveOutstanding tick is not
transactional. It called markApproved, recordAttempt and park on `this`, all
three of which are annotated @Transactional. Spring's transaction advice is
proxy-based, so every one of those self-calls bypassed it and the writes ran
with no transaction at all.

Each of the three does the same thing: re-read the authorization row by id, then
mutate and save it. Untransacted, the re-read returns a detached entity and the
read and the write are separate units — so the attempt counter can be read,
another writer can move it, and the stale value is then saved over the top. That
counter is what escalates a stuck approval to a human once the budget is spent,
and this is money-critical work: an approval that silently retries forever looks
exactly like an approval that is working.

Moved the three to a new WorkorderApprovalRecorder bean so the calls cross the
proxy. The transaction stays per-attempt rather than around the whole tick, on
purpose: one vendor refusing must not roll back attempts already counted for the
other workorders in the same batch.

queueApproval stays on the approver — it is called from the completion event
listener, not from inside this class, so its proxy was always in play. Sonar
agrees; it never flagged that one.

The 1000-character review-reason truncation moves with the writers and is now a
named constant rather than a pair of literals.
